+++ tiger-proxy/src/main/java/de/gematik/test/tiger/proxy/TigerProxy.java
@@ -33,6 +33,7 @@ import de.gematik.test.tiger.mockserver.configuration.MockServerConfiguration;
 import de.gematik.test.tiger.mockserver.mock.Expectation;
 import de.gematik.test.tiger.mockserver.netty.MockServer;
 import de.gematik.test.tiger.mockserver.proxyconfiguration.ProxyConfiguration;
+import de.gematik.test.tiger.mockserver.socket.tls.KeyAndCertificateFactory;
 import de.gematik.test.tiger.mockserver.socket.tls.KeyAndCertificateFactorySupplier;
 import de.gematik.test.tiger.proxy.client.TigerRemoteProxyClient;
 import de.gematik.test.tiger.proxy.configuration.ProxyConfigurationConverter;
@@ -43,8 +44,9 @@ import de.gematik.test.tiger.proxy.handler.BinaryExchangeHandler;
 import de.gematik.test.tiger.proxy.handler.ForwardAllCallback;
 import de.gematik.test.tiger.proxy.handler.ForwardProxyCallback;
 import de.gematik.test.tiger.proxy.handler.ReverseProxyCallback;
-import de.gematik.test.tiger.proxy.tls.DynamicTigerKeyAndCertificateFactory;
-import de.gematik.test.tiger.proxy.tls.StaticTigerKeyAndCertificateFactory;
+import de.gematik.test.tiger.proxy.tls.CombinedKeyAndCertificateFactory;
+import de.gematik.test.tiger.proxy.tls.DynamicKeyAndCertificateFactory;
+import de.gematik.test.tiger.proxy.tls.StaticKeyAndCertificateFactory;
 import io.netty.handler.ssl.SslProvider;
 import jakarta.annotation.PreDestroy;
 import java.net.*;
@@ -68,7 +70,7 @@ import org.bouncycastle.jsse.provider.BouncyCastleJsseProvider;
 public class TigerProxy extends AbstractTigerProxy implements AutoCloseable, RbelMessagesSupplier {
 
   private static final String CA_CERT_ALIAS = "caCert";
-  private final List<DynamicTigerKeyAndCertificateFactory> tlsFactories = new ArrayList<>();
+  private final List<KeyAndCertificateFactory> tlsFactories = new ArrayList<>();
   private final List<Consumer<Throwable>> exceptionListeners = new ArrayList<>();
   @Getter private final MockServerToRbelConverter mockServerToRbelConverter;
   private final Map<String, TigerRoute> tigerRouteMap = new HashMap<>();
@@ -195,13 +197,15 @@ public class TigerProxy extends AbstractTigerProxy implements AutoCloseable, Rbe
     mockServerConfiguration.enableTlsTermination(
         getTigerProxyConfiguration().isActivateTlsTermination());
 
-    final Optional<ProxyConfiguration> proxyConfiguration = ProxyConfigurationConverter.convertForwardProxyConfigurationToMockServerConfiguration(
-      getTigerProxyConfiguration());
+    final Optional<ProxyConfiguration> proxyConfiguration =
+        ProxyConfigurationConverter.convertForwardProxyConfigurationToMockServerConfiguration(
+            getTigerProxyConfiguration());
     outputForwardProxyConfigLogs(proxyConfiguration);
     proxyConfiguration.ifPresent(mockServerConfiguration::proxyConfiguration);
 
     if (getTigerProxyConfiguration().getDirectReverseProxy() == null) {
-      mockServer = new MockServer(mockServerConfiguration, getTigerProxyConfiguration().getPortAsArray());
+      mockServer =
+          new MockServer(mockServerConfiguration, getTigerProxyConfiguration().getPortAsArray());
     } else {
       mockServer = spawnDirectInverseTigerProxy(mockServerConfiguration);
     }
@@ -217,22 +221,20 @@ public class TigerProxy extends AbstractTigerProxy implements AutoCloseable, Rbe
     }
   }
 
-  private MockServer spawnDirectInverseTigerProxy(
-      MockServerConfiguration mockServerConfiguration) {
+  private MockServer spawnDirectInverseTigerProxy(MockServerConfiguration mockServerConfiguration) {
     mockServerConfiguration.binaryProxyListener(new BinaryExchangeHandler(this));
     if (mockServerConfiguration.proxyConfiguration() != null) {
       throw new TigerProxyStartupException(
           "DirectForwardProxy configured with additional forwardProxy: Not possible! (forwardProxy"
               + " is always HTTP!)");
     }
-    mockServerConfiguration.directForwarding(InetSocketAddress.createUnresolved(
-        getTigerProxyConfiguration().getDirectReverseProxy().getHostname(),
-        getTigerProxyConfiguration().getDirectReverseProxy().getPort()));
+    mockServerConfiguration.directForwarding(
+        InetSocketAddress.createUnresolved(
+            getTigerProxyConfiguration().getDirectReverseProxy().getHostname(),
+            getTigerProxyConfiguration().getDirectReverseProxy().getPort()));
 
     MockServer newMockServer =
-        new MockServer(
-            mockServerConfiguration,
-            getTigerProxyConfiguration().getPortAsArray());
+        new MockServer(mockServerConfiguration, getTigerProxyConfiguration().getPortAsArray());
     addReverseProxyRouteIfNotPresent();
 
     getRbelLogger()
@@ -300,31 +302,34 @@ public class TigerProxy extends AbstractTigerProxy implements AutoCloseable, Rbe
     return (isServerInstance, mockServerConfiguration) -> {
       val tlsConfiguration = Optional.ofNullable(getTigerProxyConfiguration().getTls());
       if (isServerInstance) {
+        KeyAndCertificateFactory serverIdentityFactory =
+            new DynamicKeyAndCertificateFactory(
+                getTigerProxyConfiguration(),
+                determineServerRootCa()
+                    .orElseThrow(
+                        () ->
+                            new TigerProxyStartupException(
+                                "Unable to determine server root CA for TigerProxy TLS Server Identity")),
+                mockServerConfiguration);
         if (tlsConfiguration.map(TigerTlsConfiguration::getServerIdentity).isPresent()) {
-          return new StaticTigerKeyAndCertificateFactory(
-              tlsConfiguration.get().getServerIdentity());
+          serverIdentityFactory =
+              new CombinedKeyAndCertificateFactory(new StaticKeyAndCertificateFactory(
+                List.of(tlsConfiguration.get().getServerIdentity())), serverIdentityFactory);
         } else if (tlsConfiguration.map(TigerTlsConfiguration::getServerIdentities).isPresent()) {
-          return new StaticTigerKeyAndCertificateFactory(
-              tlsConfiguration.get().getServerIdentities().stream()
-                  .map(TigerPkiIdentity.class::cast)
-                  .toList());
-        } else {
-          final DynamicTigerKeyAndCertificateFactory dynamicTigerKeyAndCertificateFactory =
-              new DynamicTigerKeyAndCertificateFactory(
-                  getTigerProxyConfiguration(),
-                  determineServerRootCa()
-                      .orElseThrow(
-                          () -> new TigerProxyStartupException("Unrecoverable TLS startup state")),
-                  mockServerConfiguration);
-          this.tlsFactories.add(dynamicTigerKeyAndCertificateFactory);
-          return dynamicTigerKeyAndCertificateFactory;
+          serverIdentityFactory =
+            new CombinedKeyAndCertificateFactory(new StaticKeyAndCertificateFactory(
+                  tlsConfiguration.get().getServerIdentities().stream()
+                      .map(TigerPkiIdentity.class::cast)
+                      .toList()), serverIdentityFactory);
         }
+        this.tlsFactories.add(serverIdentityFactory);
+        return serverIdentityFactory;
       } else {
         if (tlsConfiguration.map(TigerTlsConfiguration::getForwardMutualTlsIdentity).isPresent()) {
-          return new StaticTigerKeyAndCertificateFactory(
-              tlsConfiguration.get().getForwardMutualTlsIdentity());
+          return new StaticKeyAndCertificateFactory(
+              Collections.singletonList(tlsConfiguration.get().getForwardMutualTlsIdentity()));
         } else {
-          return new DynamicTigerKeyAndCertificateFactory(
+          return new DynamicKeyAndCertificateFactory(
               getTigerProxyConfiguration(),
               new TigerPkiIdentity(
                   "CertificateAuthorityCertificate.pem;CertificateAuthorityPrivateKey.pem;PKCS1"),
@@ -496,8 +501,10 @@ public class TigerProxy extends AbstractTigerProxy implements AutoCloseable, Rbe
     Objects.requireNonNull(getTigerProxyConfiguration().getTls())
         .setAlternativeNames(newAlternativeNames);
 
-    for (final DynamicTigerKeyAndCertificateFactory tlsFactory : tlsFactories) {
-      tlsFactory.addAlternativeName(host);
+    for (KeyAndCertificateFactory tlsFactory : tlsFactories) {
+      if (tlsFactory instanceof DynamicKeyAndCertificateFactory) {
+        ((DynamicKeyAndCertificateFactory) tlsFactory).addAlternativeName(host);
+      }
     }
   }