+++ tiger-proxy/src/test/java/de/gematik/test/tiger/proxy/TestTigerProxyTls.java
@@ -74,6 +74,7 @@ import org.bouncycastle.tls.DefaultTlsClient;
 import org.bouncycastle.tls.ServerOnlyTlsAuthentication;
 import org.bouncycastle.tls.TlsAuthentication;
 import org.bouncycastle.tls.TlsClientProtocol;
+import org.bouncycastle.tls.TlsException;
 import org.bouncycastle.tls.TlsServerCertificate;
 import org.bouncycastle.tls.crypto.impl.bc.BcTlsCrypto;
 import org.jetbrains.annotations.NotNull;
@@ -351,7 +352,7 @@ class TestTigerProxyTls extends AbstractTigerProxyTest {
 
   @NotNull
   private static List<String> splitDn(String s1) {
-    return Stream.of(s1.split(",")).map(String::trim).collect(Collectors.toList());
+    return Stream.of(s1.split(",")).map(String::trim).toList();
   }
 
   @ParameterizedTest
@@ -475,7 +476,7 @@ class TestTigerProxyTls extends AbstractTigerProxyTest {
                 .getStatus())
         .isEqualTo(666);
     assertThatThrownBy(() -> unirestInstance.get("https://falsche-url/foobar").asString())
-        .hasCauseInstanceOf(SSLPeerUnverifiedException.class);
+        .hasCauseInstanceOf(TlsException.class);
   }
 
   @SneakyThrows
@@ -830,6 +831,9 @@ class TestTigerProxyTls extends AbstractTigerProxyTest {
         TigerProxyConfiguration.builder()
             .tls(
                 TigerTlsConfiguration.builder()
+                    .serverRootCa(
+                        new TigerConfigurationPkiIdentity(
+                            "src/test/resources/selfSignedCa/rootCa.p12;00"))
                     .serverIdentities(
                         List.of(
                             new TigerConfigurationPkiIdentity("src/test/resources/rsa.p12;00"),
@@ -847,39 +851,42 @@ class TestTigerProxyTls extends AbstractTigerProxyTest {
                         .from("/")
                         .to("http://localhost:" + fakeBackendServerPort + "/deep/foobar")
                         .hosts(List.of("CGMAG-IM-FDSIM.ts-ttcn3.sig-test.telematik-test"))
+                        .build(),
+                    TigerRoute.builder()
+                        .from("/")
+                        .to("http://localhost:" + fakeBackendServerPort + "/foobar")
+                        .hosts(List.of("some.other.server"))
                         .build()))
             .build());
-    final UnirestInstance kon2Unirest = Unirest.spawnInstance();
-    kon2Unirest
-        .config()
-        .sslContext(
-            buildSslContextTrustingOnly(
-                new TigerConfigurationPkiIdentity("src/test/resources/rsa.p12;00")));
-    kon2Unirest
-        .get("https://127.0.0.1:" + tigerProxy.getProxyPort() + "/")
-        .header("host", "kon-instanz2.titus.ti-dienste.de")
-        .asJson();
-    final UnirestInstance cmagUnirest = Unirest.spawnInstance();
-    cmagUnirest
-        .config()
-        .sslContext(
-            buildSslContextTrustingOnly(
-                new TigerConfigurationPkiIdentity(
-                    "src/test/resources/eccServerCertificate.p12;00")));
-    cmagUnirest
-        .get("https://127.0.0.1:" + tigerProxy.getProxyPort() + "/")
-        .header("host", "CGMAG-IM-FDSIM.ts-ttcn3.sig-test.telematik-test")
-        .asJson();
 
-    awaitMessagesInTiger(4);
+    executeRequestToPathWhileOnlyTrusting(
+        "kon-instanz2.titus.ti-dienste.de", "src/test/resources/rsa.p12;00");
+    executeRequestToPathWhileOnlyTrusting(
+        "CGMAG-IM-FDSIM.ts-ttcn3.sig-test.telematik-test",
+        "src/test/resources/eccServerCertificate.p12;00");
+    executeRequestToPathWhileOnlyTrusting(
+        "some.other.server",
+        "src/test/resources/selfSignedCa/rootCa.p12;00");
 
     assertThatThrownBy(
             () ->
-                kon2Unirest
-                    .get("https://127.0.0.1:" + tigerProxy.getProxyPort() + "/")
-                    .header("host", "CGMAG-IM-FDSIM.ts-ttcn3.sig-test.telematik-test")
-                    .asString())
+              executeRequestToPathWhileOnlyTrusting(
+                "CGMAG-IM-FDSIM.ts-ttcn3.sig-test.telematik-test", "src/test/resources/rsa.p12;00"))
         .isInstanceOf(UnirestException.class)
         .hasRootCauseInstanceOf(CertPathBuilderException.class);
+  }
+
+  private void executeRequestToPathWhileOnlyTrusting(
+      String host, String fileLoadingInformation) {
+    try (UnirestInstance unirestInstance = Unirest.spawnInstance()) {
+      unirestInstance
+          .config()
+          .sslContext(
+              buildSslContextTrustingOnly(new TigerConfigurationPkiIdentity(fileLoadingInformation)));
+      unirestInstance
+          .get("https://127.0.0.1:" + tigerProxy.getProxyPort() + "/")
+          .header("host", host)
+          .asJson();
+    }
   }
 }